+++ R/generator.R
@@ -118,7 +118,14 @@ powerRelationGenerator <- function(coalitions, startWithLinearOrder = FALSE) {
   partCum <- c(0, cumsum(part))
   permsI <- 0
 
+  done <- FALSE
+
   nextPartition <- function() {
+    if(compI >= ncol(compositions)) {
+      done <<- TRUE
+      return()
+    }
+
     compI <<- compI + 1
     part <<- Filter(function(x) x != 0, compositions[,compI])
     perms <<- partitions::multinomial(part)
@@ -129,10 +136,10 @@ powerRelationGenerator <- function(coalitions, startWithLinearOrder = FALSE) {
 
   function() {
     if(permsI >= ncol(perms)) {
-      if(compI >= ncol(compositions))
-        return(NULL)
-
       nextPartition()
+    }
+    if(done) {
+      return(NULL)
     }
 
     permsI <<- permsI + 1
@@ -161,7 +168,7 @@ powerRelationGenerator <- function(coalitions, startWithLinearOrder = FALSE) {
       eqs,
       elements = elements,
       coalitionLookup = function(v) coalitionLookup[[toKey(v)]],
-      elementLookup = function(e) elementLookup[[paste(e)]]
+      elementLookup = function(e) stop('not implemented')# elementLookup[[paste(e)]]
     )
   }
 }