+++ NEWS.md
@@ -10,6 +10,7 @@ This major update brings a lot of breaking changes that are hopefully justified.
   * Added `$elementLookup()` to make it easy (and fast) to find coalitions that an element appears in
   * Changed coalitions to no longer be `sets::set` objects, but just normal vectors. This allows for one element to appear multiple times in the same coalition, which is more consistent with the ability to have a coalition appear multiple times in a `PowerRealtion` object.
 * Added `PowerRelation()` and `as.PowerRelation()`
+* Added `SocialRankingSolution()`
 * Added `coalitionLookup()` and `elementLookup()`
 * Added `L1Scores()` and `L1Ranking()`, along with its aliases `lexcel1...()`
 * Added `makePowerRelationMonotonic()` and `makePowerRelationTotal()`